separate configs, extend manager interface

+++ pkg/provisioner/deletions.go
@@ -5,7 +5,9 @@ import (
 	"os"
 
 	"github.com/martinohmann/kubernetes-cluster-manager/pkg/api"
+	"github.com/martinohmann/kubernetes-cluster-manager/pkg/config"
 	"github.com/martinohmann/kubernetes-cluster-manager/pkg/kubernetes"
+	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 )
 
@@ -22,8 +24,13 @@ func loadDeletions(filename string) (*api.Deletions, error) {
 	return &deletions, err
 }
 
-func processResourceDeletions(kubectl *kubernetes.Kubectl, deletions []*api.Deletion) error {
+func processResourceDeletions(cfg *config.Config, kubectl *kubernetes.Kubectl, deletions []*api.Deletion) error {
 	for _, deletion := range deletions {
+		if cfg.DryRun {
+			log.Warnf("Would delete the following resource:\n%s", deletion)
+			continue
+		}
+
 		if err := kubectl.DeleteResource(deletion); err != nil {
 			return err
 		}
